Cleaned up the Entities tab in the F4 Menu
The entities tab looked like a bit of a mess, so I decided to expand on
it. Ammo and weapons have been split to their own collapsible lists.
Shipments now have their own list (to make things less crowded) and the
other generic entities such as printers and other custom entities will
go into the existing entities list.

In addition, I reused the concept of hiding the vehicles list when its
empty, to also hide the other 4 lists when (and if) they are empty.
Chances are that this isn't required for anything other than the
shipments and weapons list, however I felt it was needed.