Added Dynamic Fall Damage configuration
Too lazy to detail how to use it as its self explanatory beyond this,
but GM.Config.falldamageamount is your static damage, setting to 0
disables fall damage if you aren't using dynamic damage... setting to
negative should heal in theory but dont do it cause its stupid.

GM.Config.falldamagedamper is your damper on dynamic fall damage. The
higher the number, the more resistant you are to fall damage. The lower,
the less resistant.  This is used in division so don't set it to 0 or
you will destroy time and space.

GM.Config.realisticfalldamage is true by default. Set to false for
static fall damage. mp_falldamage being set to 1 will override this.

Redundancy is in place for idiots who dont update their config.