Feature: Colors on Cheques (explained)
I changed the cheques so that they are colored for various
circumstances.

1. If you own the cheque and its for you, its yellow.
2. If you own the cheque and its for someone else, its blue.
3. If you dont own the cheque and its for you, its green.
4. If you dont own the cheque and its not for you, its red.
5. For any other circumstance not forseen, its white (redundancy
purposes, i dont think we will be needing it)

Only semi-tested. Circumstances 1 and 2 are confirmed to work, but ULX
wouldn't allow me to cexec a bot to spawn a cheque to me, so i couldn't
test the rest.