Raised FAdmin Messaging box.
Because when it was set to 20 pixels above the bottom of the screen, it
was partially cut off, something ive hated for months. 33 seems perfect
in my opinion.

Heres a before and after.

Before: http://puu.sh/45jbF.jpg

After: http://puu.sh/45jyP.jpg

FYI this has been an issue for me in DarkRP since GMOD 12. I'm glad I
finally found where to fix it.