Blind improvement to 18469fb
oogityboogity pointed out how a backslash could bypass this patch, but
there was already code to replace backslashes with forward slashes, so I
simply moved the patch down past that conversion, hoping for the best.

I haven't tested it, but if im reading it right, it will treat the
backslashes as forward slashes, and would then be picked up by the
patch.