Added MIT License
Fixes https://github.com/FPtje/DarkRP/issues/1374
Note: this license applies to all code written by Falco Peijnenburg (currently nicknamed FPtje)
Falco is not the original author of DarkRP. However, as the developer of DarkRP since 2008,
it is believed that most, if not all code has been replaced by his own.

Counterclaims on ownership of code can be discussed on the bug tracker.