Fixed QAC banning innocent players when clientside Lua is allowed.
A while ago Zer0thefallen and I had a chat about this. The problem is that his anticheat bans players who run scripts even when running scripts is /explicitly/ allowed. When clientside scripts are explicitly allowed, running them is not a crime. People who run clientside scripts with allowcslua 1 are innocent. In contrast, people who run clientside scripts with allowcslua 0 are /not/ innocent. Zer0 and I had agreed that this was the case and he promised that he would fix it in QAC. This either didn't happen or he reverted the change, letting his bullshit anticheat permaban innocent people.

I cannot let that happen, simple as that.