Added Client-Side Script Notification
I'm sure I am going to get alot of shit for this, but really its kindof a dick move to enable this by default and not mention the gamemode is changing a convar the server owner may have thought they disabled in the server's config.

As for the whole QAC thing, to be honest if the server owner installed QAC they more than likely wanted to prevent people from running client-side scripts. They probably just didn't expect the gamemode to override their existing server configuration, which is why people ended up getting banned in the first place.