Easier Moneyprinter Removal
This functionality can be used by players to quickly and efficiently remove all of their printers when they don't need them anymore. As a positive side effect this can also be used by experienced and trustworthy players to restart the server (by crashing) or to kick players in case there is a lot of lag going on or someone is failRPing and the server admins aren't available, thus improving the overall RP experience for the rest of the players online.