encourage robbery & combat the bourgeoisie
The rich get poorer and the poor must turn to crime to make a living, causing more police to join the force. Meanwhile, the police themselves as well as the rest of the bourgeoisie slowly become poorer, causing everyone to look towards more illegal activities. This causes major collapses in society which eventually lead to something like the French revolution.