Fixed issue with overlapping group chats.
Imagine two overlapping group chats: one with teams A and B,
and another with teams A and C.

When a player in team A says something in the group,
EITHER the players with team B OR the players in C would
receive the message, but never both. That is fixed now.

Note: When someone with team B sends a message, players in
team C will still not receive the message. I.e. group chats
are not transitive.

Fixes #2039