Allowing Admins Lockdown/UnLockdown
Granting ability to Admins to lockdown and unlockdown.
I would add if to the CAMI privilege of DarkRP_AdminCommands but I was unsure if what I have will work. What I had is this     CAMI.PlayerHasAccess(ply, "DarkRP_AdminCommands", function(access) 
     if not access then return end 
)
If it does, I will change the Commit (if possible).