Way better error messages when the database fucked
Should help with #2459 in the sense that the error messages will now
clearly tell what's going on and who is to blame.

Also, in combination with the MySQLite update, corrupt database error
messages were previously "hidden" because the error was the same in every
query. With that fixed, it should be very clear what's going on when the
database is fucked.