FAdmin CAMI: non-default inheritance and privs
See #2544 for the non-default inheritance issue.

While working on this I notice that the inherited usergroup referenced the
parent's privileges table. This might result in a bug that cause parents
might also get privileges when given to the child until server restart. I
haven't tested it, though. It's a preventative measure.