Updated spawning of entities to prevent unstucking
This pull request fixes a lot of situations, in which entities are spawned in a wall or weapons get lost because dropping in the floor.

It uses the same functions as sandbox to ensure everything spawns properly on solid surface.

If an entity has an spawn function written onto it, it will not try to unstuck it, just like in the sandbox gamemode.