Database: Half decent migration system + 20190914 migration
See #2947

The previous migration system was a half assed combination of using a DB version
and just checking whether tables and stuff, and migrate based on that.

The code used to always run migrations, with the knowledge that they would be
idempotent. Now the migrations are not performed unless really necessary. This
saves unnecessary code from being run (@McNuggets1), specifically it saves the
running of queries at startup.