Optimise server's side of entityvars.lua
- The client already has this optimisation in place and this results in much faster access time
FProfiler has shown, that "getDarkRPVar" is one of the most expensive bottlenecks

- Renamed unnamed hook
This hook fooled me a few times already into searching for it, because I didn't know what it did, while working with hook.GetTable()[index] to find a hook returning something

This PR resolves this.

Fixing this problem ALONE is not possible without modifying DarkRP core files.