Move tremor hook.Add to the file scope
Moving createTremor to inside InitPostEntity is a reasonable fix, but there is 0 reason for the cleanup hook to be inside of another function. Map cleanup should not be done before InitPostEntity and this was only implemented this way because some bad addons call game.CleanUpMap() before entities have initialised which is what should be fixed. There are many other things that break (not necessarily in DarkRP) from this incorrect call and this is a bandaid on a gash to only cause InitPostEntity errors to have even greater detrimental effects.